+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped `glam` to 0.30. [Changelog](https://github.com/bitshifter/glam-rs/blob/main/CHANGELOG.md#0300---2025-02-18).
 - Bumped `wasmtime` dependency to 29.x.
+- Shuffled trait bounds between `IntUnit` and `IntScalar` to support mapping integers to
+  their unsigned counterparts (needed for `manhattan_distance()` etc.).
 
 ## [0.15.0] - 2024-10-12